+++ source/standalone/workflows/skrl/play.py
@@ -25,6 +25,13 @@ parser.add_argument("--cpu", action="store_true", default=False, help="Use CPU p
 parser.add_argument("--num_envs", type=int, default=None, help="Number of environments to simulate.")
 parser.add_argument("--task", type=str, default=None, help="Name of the task.")
 parser.add_argument("--checkpoint", type=str, default=None, help="Path to model checkpoint.")
+parser.add_argument(
+    "--framework",
+    type=str,
+    default="torch",
+    choices=["torch", "jax.jax", "jax.numpy"],
+    help="Deep learning framework to use.",
+)
 args_cli = parser.parse_args()
 
 # launch the simulator
@@ -36,13 +43,24 @@ simulation_app = SimulationApp(config)
 
 import gym
 
-from skrl.agents.torch.ppo import PPO, PPO_DEFAULT_CONFIG
-from skrl.utils.model_instantiators import deterministic_model, gaussian_model, shared_model
+if args_cli.framework.startswith("torch"):
+    from skrl.agents.torch.ppo import PPO, PPO_DEFAULT_CONFIG
+    from skrl.utils.model_instantiators.torch import deterministic_model, gaussian_model, shared_model
+elif args_cli.framework.startswith("jax"):
+    import jax  # fmt:skip
+    jax.Device = jax.xla.Device  # for Isaac Sim 2022.2.1 or earlier
+
+    from skrl.agents.jax.ppo import PPO, PPO_DEFAULT_CONFIG
+    from skrl.utils.model_instantiators.jax import deterministic_model, gaussian_model
 
 import omni.isaac.contrib_envs  # noqa: F401
 import omni.isaac.orbit_envs  # noqa: F401
 from omni.isaac.orbit_envs.utils import get_checkpoint_path, parse_env_cfg
-from omni.isaac.orbit_envs.utils.wrappers.skrl import SkrlVecEnvWrapper
+
+if args_cli.framework == "torch":
+    from omni.isaac.orbit_envs.utils.wrappers.skrl import SkrlTorchVecEnvWrapper as SkrlVecEnvWrapper
+elif args_cli.framework.startswith("jax"):
+    from omni.isaac.orbit_envs.utils.wrappers.skrl import SkrlJaxVecEnvWrapper as SkrlVecEnvWrapper
 
 from config import convert_skrl_cfg, parse_skrl_cfg
 
@@ -59,21 +77,24 @@ def main():
     env = SkrlVecEnvWrapper(env)  # same as: `wrap_env(env, wrapper="isaac-orbit")`
 
     # instantiate models using skrl model instantiator utility
-    # https://skrl.readthedocs.io/en/latest/modules/skrl.utils.model_instantiators.html
+    # https://skrl.readthedocs.io/en/latest/api/utils/model_instantiators.html
     models = {}
+    # force separated models for jax
+    if args_cli.framework.startswith("jax"):
+        experiment_cfg["models"]["separate"] = True
     # non-shared models
     if experiment_cfg["models"]["separate"]:
         models["policy"] = gaussian_model(
             observation_space=env.observation_space,
             action_space=env.action_space,
             device=env.device,
-            **convert_skrl_cfg(experiment_cfg["models"]["policy"]),
+            **convert_skrl_cfg(experiment_cfg["models"]["policy"], args_cli.framework),
         )
         models["value"] = deterministic_model(
             observation_space=env.observation_space,
             action_space=env.action_space,
             device=env.device,
-            **convert_skrl_cfg(experiment_cfg["models"]["value"]),
+            **convert_skrl_cfg(experiment_cfg["models"]["value"], args_cli.framework),
         )
     # shared models
     else:
@@ -84,17 +105,21 @@ def main():
             structure=None,
             roles=["policy", "value"],
             parameters=[
-                convert_skrl_cfg(experiment_cfg["models"]["policy"]),
-                convert_skrl_cfg(experiment_cfg["models"]["value"]),
+                convert_skrl_cfg(experiment_cfg["models"]["policy"], args_cli.framework),
+                convert_skrl_cfg(experiment_cfg["models"]["value"], args_cli.framework),
             ],
         )
         models["value"] = models["policy"]
+    # initialize jax models' state dict
+    if args_cli.framework.startswith("jax"):
+        for role, model in models.items():
+            model.init_state_dict(role)
 
     # configure and instantiate PPO agent
-    # https://skrl.readthedocs.io/en/latest/modules/skrl.agents.ppo.html
+    # https://skrl.readthedocs.io/en/latest/api/agents/ppo.html
     agent_cfg = PPO_DEFAULT_CONFIG.copy()
     experiment_cfg["agent"]["rewards_shaper"] = None  # avoid 'dictionary changed size during iteration'
-    agent_cfg.update(convert_skrl_cfg(experiment_cfg["agent"]))
+    agent_cfg.update(convert_skrl_cfg(experiment_cfg["agent"], args_cli.framework))
 
     agent_cfg["state_preprocessor_kwargs"].update({"size": env.observation_space, "device": env.device})
     agent_cfg["value_preprocessor_kwargs"].update({"size": 1, "device": env.device})